requirements.txt
elevenlabs
dotenv
aiohttp
google-genai
PyYAML
sounddevice
soundfile
numpy

